requirements.txt
dnspython
rich
typer
aiohttp
cryptography